alestar

+++ app/Http/Controllers/Admin/PlanController.php
@@ -58,7 +58,7 @@ class PlanController extends Controller
     public function show($url)
     {
         $plan = $this->repository->where('url', $url)->first();
-        
+
         if(!$plan)
             return redirect()->back();
 
@@ -112,14 +112,18 @@ class PlanController extends Controller
      */
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository->with('details')->where('url', $url)->first();
 
         if(!$plan)
             return redirect()->back();
 
+        if ($plan->details->count() > 0) {
+            return redirect()->back()->with('error', 'Existem detalhes vinculados a esse plano, por tanto não pode deletar');
+        }
+
         $plan->delete();
 
-        return redirect()->route('plans.index');
+        return redirect()->route('plans.index')->with('message', 'Registro deletado com sucesso.');
     }
 
     public function search(Request $request)
